LayoutTests/imported/w3c/web-platform-tests/workers/name-property-expected.txt

CONSOLE MESSAGE: line 20: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = Error in remote http://localhost:8800/workers/support/name-as-accidental-global.js: ReferenceError: Can't find variable: test

FAIL Test the name property of shared and dedicated workers via the name constructor option ReferenceError: Can't find variable: SharedWorker
FAIL name property value for DedicatedWorkerGlobalScope assert_true: property exists on the global expected true got false
PASS name property is replaceable for DedicatedWorkerGlobalScope 
